+++ tests/common/mod.rs
@@ -7,5 +7,6 @@ pub use assert_output::TEMPDIR_PLACEHOLDER;
 pub use assert_output::assert_output_contains;
 pub use assert_output::replace_tempdir;
 pub use avro::assert_avro_codec;
+pub use parquet::assert_parquet_codec;
 pub use parquet::assert_valid_parquet_file;
 pub use row_count::get_row_count;